move function to init index maps to BGL
