+++ BGL/include/CGAL/boost/graph/properties.h
@@ -23,6 +23,7 @@
 
 #include <CGAL/property_map.h>
 #include <boost/graph/properties.hpp>
+#include <boost/graph/graph_traits.hpp>
 
 #include <CGAL/basic.h>
 #include <string>
@@ -104,5 +105,128 @@ using boost::face_external_index_t;
 using boost::face_external_index;
 using boost::vertex_property_t;
 } // CGAL
+
+namespace CGAL{
+namespace helpers {
+
+// matches read-write property maps
+template <class PolygonMesh, class FaceIndexMap, class Tag>
+void init_face_indices(PolygonMesh& pm,
+                       FaceIndexMap& fid,
+                       boost::read_write_property_map_tag,
+                       Tag)
+{
+  typename boost::property_traits<FaceIndexMap>::value_type i = 0;
+  BOOST_FOREACH(typename boost::graph_traits<PolygonMesh>::face_descriptor fd,
+                faces(pm))
+  {
+    put(fid, fd, i);
+    ++i;
+  }
+}
+template <class PolygonMesh, class VertexIndexMap, class Tag>
+void init_vertex_indices(PolygonMesh& pm,
+                         VertexIndexMap& vid,
+                         boost::read_write_property_map_tag,
+                         Tag)
+{
+  typename boost::property_traits<VertexIndexMap>::value_type i = 0;
+  BOOST_FOREACH(typename boost::graph_traits<PolygonMesh>::vertex_descriptor vd,
+                vertices(pm))
+  {
+    put(vid, vd, i);
+    ++i;
+  }
+}
+template <class PolygonMesh, class HalfedgeIndexMap, class Tag>
+void init_halfedge_indices(PolygonMesh& pm,
+                           HalfedgeIndexMap& hid,
+                           boost::read_write_property_map_tag,
+                           Tag)
+{
+  typename boost::property_traits<HalfedgeIndexMap>::value_type i = 0;
+  BOOST_FOREACH(typename boost::graph_traits<PolygonMesh>::halfedge_descriptor hd,
+                halfedges(pm))
+  {
+    put(hid, hd, i);
+    ++i;
+  }
+}
+
+// matches mutable Lvalue property maps
+template <class PolygonMesh, class FaceIndexMap>
+void init_face_indices(PolygonMesh& pm,
+                       FaceIndexMap& fid,
+                       boost::lvalue_property_map_tag,
+                       boost::false_type)
+{
+  init_face_indices(pm, fid,
+    boost::read_write_property_map_tag(), boost::false_type());
+}
+template <class PolygonMesh, class VertexIndexMap>
+void init_vertex_indices(PolygonMesh& pm,
+                         VertexIndexMap& vid,
+                         boost::lvalue_property_map_tag,
+                         boost::false_type)
+{
+  init_vertex_indices(pm, vid,
+    boost::read_write_property_map_tag(), boost::false_type());
+}
+template <class PolygonMesh, class HalfedgeIndexMap>
+void init_halfedge_indices(PolygonMesh& pm,
+                         HalfedgeIndexMap& hid,
+                         boost::lvalue_property_map_tag,
+                         boost::false_type)
+{
+  init_halfedge_indices(pm, hid,
+    boost::read_write_property_map_tag(), boost::false_type());
+}
+
+// matches all other types of property map
+template <class PolygonMesh, class FaceIndexMap, class MapTag, class Tag>
+void init_face_indices(PolygonMesh&, FaceIndexMap, MapTag, Tag)
+{}
+template <class PolygonMesh, class VertexIndexMap, class MapTag, class Tag>
+void init_vertex_indices(PolygonMesh&, VertexIndexMap, MapTag, Tag)
+{}
+template <class PolygonMesh, class HalfedgeIndexMap, class MapTag, class Tag>
+void init_halfedge_indices(PolygonMesh&, HalfedgeIndexMap, MapTag, Tag)
+{}
+
+template <class PolygonMesh, class FaceIndexMap>
+void init_face_indices(PolygonMesh& pm, FaceIndexMap fid)
+{
+  init_face_indices(pm, fid,
+                    typename boost::property_traits<FaceIndexMap>::category(),
+                    typename boost::is_const<
+                      typename boost::remove_reference<
+                        typename boost::property_traits<FaceIndexMap>::reference
+                            >::type >::type() );
+}
+
+template <class PolygonMesh, class VertexIndexMap>
+void init_vertex_indices(PolygonMesh& pm, VertexIndexMap vid)
+{
+  init_vertex_indices(pm, vid,
+                      typename boost::property_traits<VertexIndexMap>::category(),
+                      typename boost::is_const<
+                        typename boost::remove_reference<
+                          typename boost::property_traits<VertexIndexMap>::reference
+                            >::type >::type() );
+}
+
+template <class PolygonMesh, class HalfedgeIndexMap>
+void init_halfedge_indices(PolygonMesh& pm, HalfedgeIndexMap hid)
+{
+  init_halfedge_indices(pm, hid,
+                        typename boost::property_traits<HalfedgeIndexMap>::category(),
+                        typename boost::is_const<
+                          typename boost::remove_reference<
+                            typename boost::property_traits<HalfedgeIndexMap>::reference
+                              >::type >::type() );
+}
+
+} } //end of namespace CGAL::helpers
+
 
 #endif // CGAL_BOOST_GRAPH_BGL_PROPERTIES_H
